% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binBoundary.R
\name{bin_boundaries}
\alias{bin_boundaries}
\title{Determine bin boundaries}
\usage{
bin_boundaries(data_dir, bincount = 5000)
}
\arguments{
\item{data_dir}{The directory that contains mappable region file (mappable.regions.txt) generated by simulation of reference genome.
Also contains the output files: bin boundary file (bin.boundaries.nk.sorted.txt), mappable.regions.sorted.txt, chrom.mappable.txt.}

\item{bincount}{An integer. The number of bins. Default value: 5000.}
}
\description{
Make bin boundary files: bin.boundaries.nk.sorted.txt
}
