% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binCount.R
\name{bin_counts}
\alias{bin_counts}
\title{make bin count files}
\usage{
bin_counts(SAM_dir, data_dir, cell_name, Nk = "5k")
}
\arguments{
\item{SAM_dir}{The directory that contains the sorted SAM files with duplicates removed. The SAM file names: cellname.rmdup.sam.
e.g. CJA1023.rmdup.sam
The bin count files will also be saved in SAM_dir. The output bin count file names: cellname.varbin.Nk.txt, cellname.varbin.Nk.stats.txt
e.g. CJA1023.varbin.20k.txt, CJA1023.varbin.20k.stats.txt}

\item{data_dir}{The directory that contains the bin boundary file; also contains mappable region files such as
chrom.mappable.txt, mappable.regions.sorted.txt, chrom.sizes.txt}

\item{cell_name}{The cell name.}

\item{Nk}{Default value: 5k. It denotes the number of bins. 5000 = 5k, 10000 = 10k, ...}
}
\description{
Make bin count files from SAM files
}
