% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_bins2regions}
\alias{calc_bins2regions}
\title{Converts list of bins from binning scheme to regions.}
\usage{
calc_bins2regions(gc_df, bins)
}
\description{
Converts list of bins from binning scheme to regions.
}
