% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_centroareas}
\alias{calc_centroareas}
\title{Calculates centromere regions (areas).}
\usage{
calc_centroareas(cyto, centromere = c("p11", "q11"))
}
\description{
Calculates centromere regions (areas).
}
