% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinmat.R
\name{calc_pinmat}
\alias{calc_pinmat}
\title{Select features and generate the incidence table.}
\usage{
calc_pinmat(gc_df, segment_df, homoloss = 0, dropareas = NULL, smear = 1,
  chromrange = 1:24, keepboundaries = F)
}
\arguments{
\item{gc_df}{bining schema used for the analysis}

\item{segment_df}{the breakpoint table generated by \code{segment_varbin_files}.}

\item{homoloss}{drop out boundary}

\item{dropareas}{areas of the chromosomes that should be excluded from 
further analysis (e.g. centromeres)}
}
\value{
a list of pinmat and pins objects.
        pinmat is the incidence table; pins is the bin location
}
\description{
Select features (called as pins), generate the binary matrix with rows as pins 
and columns as cells.
}
