% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinmat.R
\name{calc_ploidies}
\alias{calc_ploidies}
\title{Calculates ploidies for samples passed in \code{segment_df}.}
\usage{
calc_ploidies(gc_df, segment_df, chromrange = 1:24)
}
\arguments{
\item{gc_df}{bining schema used for the analysis}

\item{segment_df}{the breakpoint table generated by \code{segment_varbin_files}.}
}
\value{
ploidies calculated for samples in \code{segment_df}.
}
\description{
Calculates ploidies for samples passed in \code{segment_df}.
}
