% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_regions2bins}
\alias{calc_regions2bins}
\title{Converts regions to list of bins from binning scheme.}
\usage{
calc_regions2bins(gc_df, regions)
}
\description{
Converts regions to list of bins from binning scheme.
}
