% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{case_filenames}
\alias{case_filenames}
\title{Constructs names for various output files based on `output_dir` and
`casename`}
\usage{
case_filenames(output_dir, casename)
}
\description{
Constructs names for various output files based on `output_dir` and
`casename`
}
