% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcCBS_all.R
\name{cbs.segment_all}
\alias{cbs.segment_all}
\title{Generate the segmented profile for each cell.}
\usage{
cbs.segment_all(input_file_dir, Nk = "5k", gc, alpha, nperm, undo.SD,
  min.width, method = "multiplier", genome = "hg")
}
\arguments{
\item{input_file_dir}{The directory that contains the bin count files for all cells. For example, one bin count file for a cell is CJA1023.varbin.20k.txt}

\item{Nk}{Default value: 5k. It denotes the number of bins. 5000 = 5k, 10000 = 10k, ...}

\item{gc}{The GC content table.}

\item{alpha}{Parameter alpha defined in DNAcopy package.}

\item{nperm}{Parameter nperm defined in DNAcopy package.}

\item{undo.SD}{Parameter undo.SD defined in DNAcopy package.}

\item{min.width}{Parameter min.width defined in DNAcopy package.}

\item{method}{Default: "multiplier" to transform the ratio data to integer CN state. When genome is hg-dm hybrid, set method as "dmploidies" to use dm plodies as reference.}

\item{genome}{Default: "hg" (Human genome). hg-dm hybrid genome: "hgdm".}
}
\value{
The list containing seg.quantal and ratio.quantal matrix for all cells in the input_file_dir.
}
\description{
Generate the segmented profile for each cell in the input directory with CBS. (Including GC correction)
}
