% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcCBS_one.R
\name{cbs.segment_one}
\alias{cbs.segment_one}
\title{Segment the normalzied bin count profile for individual cell for GC content.}
\usage{
cbs.segment_one(bin_mat, alpha, nperm, undo.SD, min.width,
  method = "multiplier", genome = "hg", graphic = TRUE)
}
\arguments{
\item{bin_mat}{The normalized bin count files generated by gc_one.}

\item{alpha}{Parameter alpha defined in DNAcopy package.}

\item{nperm}{Parameter nperm defined in DNAcopy package.}

\item{undo.SD}{Parameter undo.SD defined in DNAcopy package.}

\item{min.width}{Parameter min.width defined in DNAcopy package.}

\item{method}{Default: "multiplier" to transform the ratio data to integer CN state. When genome is hg-dm hybrid, set method as "dmploidies" to use dm plodies as reference.}

\item{genome}{Default: "hg" (Human genome). hg-dm hybrid genome: "hgdm".}

\item{graphic}{Default: TRUE. Generate the copy number profile plot.}
}
\value{
The segmented profile for individual cell
}
\description{
Generate the segmented profile for individual cell with CBS.
}
