% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chrom_numeric}
\alias{chrom_numeric}
\title{Converts chrom name to numeric and adds `chrom.numeric` column to the 
dataframe.}
\usage{
chrom_numeric(chrom)
}
\description{
Converts chrom name to numeric and adds `chrom.numeric` column to the 
dataframe.
}
