% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{drop_areas}
\alias{drop_areas}
\title{Identify the centromere areas}
\usage{
drop_areas(cyto)
}
\arguments{
\item{cyto}{The input annotation file "cytoBandHG19.txt"}
}
\value{
A table for genomic location of the centromere areas.
}
\description{
Return the genomic location of the centromere areas, which are considered to be filtered out later.
}

