% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1.varbin.20k}
\alias{example1.varbin.20k}
\title{example1.varbin.20k}
\format{A data matirx}
\usage{
example1.varbin.20k
}
\description{
The input bin count data for one cell. It requires GC normalization and CBS segmentation.
}
\keyword{datasets}
