% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherFDR.R
\name{fdr_fisherPV}
\alias{fdr_fisherPV}
\title{Compute FDRs for Fisher's test p-values.}
\usage{
fdr_fisherPV(true_fisherPV, sim_fisherPV, cell_names, lm_max = 0.001,
  graphic = TRUE)
}
\arguments{
\item{true_fisherPV}{The Fisher's test p-values for the observation.}

\item{sim_fisherPV}{The Fisher's test p-values for the permutations.}

\item{cell_names}{A character vector. The names of cells.}

\item{lm_max}{Numeric value. Default: 0.001. The threshold parameter for the linear fit.}

\item{graphic}{Logical. If TRUE (Default), generate the CDF plots of p-values and FDR values.}
}
\value{
A list containing the matrix of the FDR values (mat_fdr) and distance matrix based on Fisher's test p-values (mat_dist).
}
\description{
Linear fit to the tail of empirical null distribution of Fisher p-values;
FDR computation: compare true to simulated CDF(empirical null).
}
