% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClones.R
\name{find_clone}
\alias{find_clone}
\title{Identify clones in hierarchical tree.}
\usage{
find_clone(hc, fdr_thresh = -2, share_min = 0.9, n_share = 3,
  bymax = TRUE, climb_from_size = 2, climb_to_share = 3, graphic = TRUE)
}
\arguments{
\item{hc}{The hclust objects with new items added generated by \code{hclust_tree}.}

\item{fdr_thresh}{FDR criterion for clone nodes. Default: -2.}

\item{share_min}{A feature is considered shared if present in share_min fraction of leaves in a node.Default: 0.90.}

\item{n_share}{Minimal number of shared features in a clone node. Default: 3.}

\item{bymax}{Logical. If TRUE (Default), use maximal of mean FDR for the node to find clones.}

\item{climb_from_size}{An integer.}

\item{climb_to_share}{An integer.}

\item{graphic}{Logical. If TRUE (Default), generate the hierarchical tree plot with hard/soft clones labeled.}
}
\value{
A hclust object.
}
\description{
Based on hierarchical clustering, identify the hard/soft clones.
}
