% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findsubclones.R
\name{find_subclone}
\alias{find_subclone}
\title{Identify subclones in hierarchical tree.}
\usage{
find_subclone(hc, pinmat, pins, min_node_size = 6, sim_round = 500,
  lm_max = 0.001, hc_method = "average", base_share = 3,
  fdr_thresh = -2, share_min = 0.9, bymax = TRUE, climb_from_size = 2,
  climb_to_share = 3, graphic = TRUE)
}
\arguments{
\item{hc}{The hclust objects with clones identified.}

\item{pinmat}{The pinmat.}

\item{pins}{The pins.}

\item{min_node_size}{An integer. Default: 6. The minimum node size for a subclone.}

\item{sim_round}{The number of permutation simulations for subclone identification. Default: 500.}

\item{lm_max}{Numeric value. Default: 0.001. The threshold parameter for the linear fit to identify subclones.}

\item{hc_method}{Default: average}

\item{base_share}{An integer. Default: 3. A balance parameter for controlling minimal number of shared features in a subclone node.}

\item{fdr_thresh}{FDR criterion for subclone nodes. Default: -2.}

\item{share_min}{A feature is considered shared if present in share_min fraction of leaves in a node.Default: 0.90.}

\item{bymax}{Logical. If TRUE (Default), use maximal of mean FDR for the node to find subclones.}

\item{climb_from_size}{An integer. Default: 2.}

\item{climb_to_share}{An integer. Default: 3.}

\item{graphic}{Logical. If TRUE (Default), generate the hierarchical tree plot with hard/soft clones labeled.}
}
\value{
A list of hclust objects for clones.
}
\description{
Based on hierarchical clustering, identify the hard/soft clones.
}
