% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_subclones.R
\name{find_subclones}
\alias{find_subclones}
\title{Identify subclones in hierarchical tree.}
\usage{
find_subclones(hc, pinmat, pins, nmin = -6, nsim = 500, lmmax = 0.001,
  hcmethod = "average", baseshare = 3, fdrthresh = -2, sharemin = 0.85,
  bymax = T, climbfromsize = 2, climbtoshare = 3, clonetype = "soft")
}
\arguments{
\item{hc}{The hclust objects with clones identified.}

\item{pinmat}{The pinmat.}

\item{pins}{The pins.}

\item{nmin}{An integer. Default: 6. The minimum node size for a subclone.}

\item{nsim}{The number of permutation simulations for subclone identification. 
Default: 500.}

\item{lmmax}{Numeric value. Default: 0.001. The threshold parameter for the 
linear fit to identify subclones.}

\item{hcmethod}{Default: average}

\item{baseshare}{An integer. Default: 3. A balance parameter for controlling 
minimal number of shared features in a subclone node.}

\item{fdrthresh}{FDR criterion for subclone nodes. Default: -2.}

\item{sharemin}{A feature is considered shared if present in sharemin fraction 
of leaves in a node.Default: 0.85.}

\item{bymax}{Logical. If TRUE (Default), use maximal of mean FDR for the node 
to find subclones.}

\item{climbfromsize}{An integer. Default: 2.}

\item{climbtoshare}{An integer. Default: 3.}

\item{clonetype}{Default: 'soft'.}
}
\value{
A list of hclust objects for clones.
}
\description{
Based on hierarchical clustering, identify the hard/soft clones.
}
