% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectpin.R
\name{findpins}
\alias{findpins}
\title{Select features and generate the incidence table.}
\usage{
findpins(breakpoint_table, smear_table)
}
\arguments{
\item{breakpoint_table}{The breakpoint_table generated by \code{preprocess_segfile}.}

\item{smear_table}{The smear_table generated by \code{findsmears}.}
}
\value{
A list of three objects. pinmat is the incidence table. pins is the bin location and sign for
the selected features (pins).
}
\description{
Select features (called as pins), generate the binary matrix with rows as pins and columns as cells.
}
