% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectpin.R
\name{findsmears}
\alias{findsmears}
\title{Generate the smear tables.}
\usage{
findsmears(breakpoint_table, smear = 2, keepboundaries = TRUE,
  mask_XY = TRUE)
}
\arguments{
\item{breakpoint_table}{The breakpoint_table generated by \code{preprocess_segfile}.}

\item{smear}{An integer. The number of spanning bins on one side. Default value: 2.}

\item{keepboundaries}{Logical. If TRUE (Default), the bins on chromosome boundaries are kept.}

\item{mask_XY}{Logical. If TRUE (Default), sex chromosomes (X, Y) are masked.}
}
\value{
smear_table. The cell names (profid), the chromosome (chrom) and bin location (bpstart, bpend)
for the extended intervals with length of 2*smear and the change direction (bpsign).
}
\description{
Given the breakpoint tables of all single cells generated by \code{preprocess_segfile},
extend the breakpoint bins into intervals containing several bins to
allow for the uncertainty inherent in segmentation.
}
