% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_fdr.R
\name{fisher_dist}
\alias{fisher_dist}
\title{Calculates a distance matrix given Fisher FDR true p-values.}
\usage{
fisher_dist(true_pv, cell_names)
}
\arguments{
\item{true_pv}{The Fisher's test p-values for the observation.}

\item{cell_names}{A character vector. The names of cells.}
}
\value{
distance matrix based on Fisher's test p-values (mat_dist).
}
\description{
Calculates a distance matrix given Fisher FDR true p-values.
}
