% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_fdr.R
\name{fisher_fdr}
\alias{fisher_fdr}
\title{Compute FDRs for Fisher's test p-values.}
\usage{
fisher_fdr(true_pv, sim_pv, cell_names, lmmax = 0.001)
}
\arguments{
\item{true_pv}{The Fisher's test p-values for the observation.}

\item{sim_pv}{The Fisher's test p-values for the permutations.}

\item{cell_names}{A character vector. The names of cells.}

\item{lmmax}{Numeric value. Default: 0.001. The threshold parameter 
for the linear fit.}
}
\value{
A list containing the matrix of the FDR values (mat_fdr)
}
\description{
Linear fit to the tail of empirical null distribution of Fisher p-values;
FDR computation: compare true to simulated CDF(empirical null).
}
