% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcCBS_one.R
\name{gc_one}
\alias{gc_one}
\title{GC normalization.}
\usage{
gc_one(bin_mat, gc)
}
\arguments{
\item{bin_mat}{The bin count files of cells. For example, one bin count file for a cell is CJA1023.varbin.20k.txt}

\item{gc}{The GC content table.}
}
\value{
The normalized bin count table based on GC content.
}
\description{
Normalize the bin count file to adjust for nonuniformity results from variation in GC content across the genome.
}
