% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclustTree.R
\name{hclust_tree}
\alias{hclust_tree}
\title{Build the hierarchical clustering tree.}
\usage{
hclust_tree(pinmat, mat_fdr, true_fisherPV, cell_names, hc_method = "average")
}
\arguments{
\item{pinmat}{The incidence table generated by \code{findpins}.}

\item{mat_fdr}{The FDR matrix generated by \code{fdr_fisherPV}}

\item{true_fisherPV}{The Fisher's test p-values for the observations.}

\item{cell_names}{A character vector. The names of cells.}

\item{hc_method}{Default: average}
}
\value{
A hclust objects with new items added.
}
\description{
Hierarchical clustering with Fisher's test p-values as distance matrix.
Also add feature coverage information for each node in the tree.
}

