% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19.chrom.sizes}
\alias{hg19.chrom.sizes}
\title{hg19.chrom.sizes}
\format{A data matirx}
\usage{
hg19.chrom.sizes
}
\description{
The file listing the sizes of the chromosomes.
}
\keyword{datasets}
