% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19.varbin.gc.5k}
\alias{hg19.varbin.gc.5k}
\title{hg19.varbin.gc.5k}
\format{A data matirx}
\usage{
hg19.varbin.gc.5k
}
\description{
The GC content for each bin (5k bins).
}
\keyword{datasets}
