% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.R
\name{output_viewer}
\alias{output_viewer}
\title{Generate files for the downstream visualization using viewer.}
\usage{
output_viewer(output_file_dir, seg.quantal, ratio.quantal, pins, pinmat,
  mat_dist, hc_clone, sub_hc_clone, subcloneTooBig = 0.8, smear = 2, study)
}
\arguments{
\item{output_file_dir}{The directory made to save the output files for the viewer.}

\item{seg.quantal}{The seg.quantal profiles.}

\item{ratio.quantal}{The ratio.quantal profiles}

\item{pins}{The pins.}

\item{pinmat}{The pinmat.}

\item{mat_dist}{The dissmilarity based on Fisher's test p-values for hierarchical clustering.}

\item{hc_clone}{The hc object with clones identified.}

\item{sub_hc_clone}{The list of hc objects in clones with subclones identified.}

\item{subcloneTooBig}{Default: 0.8}

\item{smear}{The smear parameter. Default value: 2}

\item{study}{The name of your study.}
}
\description{
Generate files for the downstream visualization using viewer.
}
