% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_annot}
\alias{preprocess_annot}
\title{Preprocess the genomic annotation file}
\usage{
preprocess_annot(annot)
}
\arguments{
\item{annot}{The input annotation file "varbin.gc.content.5k.bowtie.k50.hg19.txt".}
}
\value{
The formated annotation table.
}
\description{
Reformat the chromosome ID of the input annotation file
}

