% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_segfile}
\alias{preprocess_segfile}
\title{Preprocess the segmented copy number profiles}
\usage{
preprocess_segfile(segfile, gc, eviltwins = NULL, ploidies = TRUE)
}
\arguments{
\item{segfile}{The segmented copy number profiles of some single cells.}

\item{gc}{The GC content table.}

\item{eviltwins}{Bad cells. NULL or a vector of cell names.}

\item{ploidies}{Logical. If TRUE (Default), the ploidies and homoloss information are taken into consideration.}
}
\value{
A list for two tables: breakpoint_table and ploidies_table. Each row in breakpoint_table correspond
to one segment divided by breakpoints in one cell. The columns in breakpoint_table: cell names (profid),
chromosome information(chrom, chromstart, chromend), absolute genomic location (abstart, absend), the bin location
(segstarts, segends), count of bins in the segement (segbins) and the copy number value for the segment (segvals, cvals).
Note: cvals are derived by incoporating ploidies.
}
\description{
Input the segmented copy number profiles of some single cells,
generate the breakpoint table which provides the genomic location
and segmented copy number values for all cells.
}
