% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{sgains_pipeline}
\alias{sgains_pipeline}
\title{Integration to 'sGAINS' tool.}
\usage{
sgains_pipeline(scgv_dir, case_name, varbin_dir, varbin_suffix, bins_boundaries,
  cytoband, badbins = NULL, nsim = 150, sharemin = 0.85)
}
\arguments{
\item{scgv_dir}{directory where the results of the analysis should be stored}

\item{case_name}{name of the case to be used for storing results of the analysis}

\item{varbin_dir}{directory where are located results from 'varbin' step of sGAINS}

\item{varbin_suffix}{common suffix for files produced from 'varbin' step of sGAINS}

\item{bins_boundaries}{file name for binning scheme to use in the analysis}

\item{cytoband}{file name where is located the description of cyto bands for
the version of genome we are using}

\item{badbins}{a file name were the defenition of bad bins for the specified
binning schema could be found}

\item{nsim}{number of simulations to run for calculating simulated FDR distribution}

\item{sharemin}{a feature is considered shared if present in sharemin fraction 
of leaves in a node}
}
\description{
This function is called by sGAINS tools to perform final step in preparation
of results.
}
