% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfisher.R
\name{sim_fisher_wrapper}
\alias{sim_fisher_wrapper}
\title{Simulate the Fisher's test p-values.}
\usage{
sim_fisher_wrapper(pinmat_df, pins_df, njobs = NULL, nsim = 150,
  nsweep = 200, seedme = 123)
}
\arguments{
\item{pinmat_df}{The incidence table generated by \code{findpins}.}

\item{pins_df}{The pin generated by \code{findpins}. The bin information for
the selected feature set.}

\item{nsim}{the number of permutations/simulations. Default value: 150.}
}
\value{
a list of two numeric vector objects. The Fisher's test p-values
        for the observation (true) and for the permutations (sim).
}
\description{
Given the incidence table for selected features (i.e. pinmat generated by
\code{calc_pins}), computes the Fisher's test p-values for pairwise comparisons.
Also perform permutations on the incidence table and compute a set of
Fisher's test p-values for each permutation.
}
