% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFisherWrapper.R
\name{simFisher_parallel}
\alias{simFisher_parallel}
\title{Simulate the Fisher's test p-values.}
\usage{
simFisher_parallel(pins, pinmat, sim_round = 500)
}
\arguments{
\item{pins}{The pin generated by \code{findpins}. The bin information for the selected feature set.}

\item{pinmat}{The incidence table generated by \code{findpins}.}

\item{sim_round}{An integer. The number of permutations/simulations. Default value: 500.}
}
\value{
A list of two numeric vector objects. The Fisher's test p-values for the observation (true_fisherPV) and for
the permutations (sim_fisherPV).
}
\description{
Given the incidence table for selected features (i.e. pinmat generated by \code{findpins}),
compute the Fisher's test p-values for pairwise comparisons. Also perform permutations on the incidence table and
compute a set of Fisher's test p-values for each permutation.
}
