% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tree_py}
\alias{tree_py}
\title{Builds HC tree representation based on the distance matrix
computed by \code{fisher_dist}
export}
\usage{
tree_py(mdist, method, metric = "euclidean")
}
\description{
Builds HC tree representation based on the distance matrix
computed by \code{fisher_dist}
export
}
