% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{varbin_input_files}
\alias{varbin_input_files}
\title{Collects all bin count files from given directory}
\usage{
varbin_input_files(input_file_dir, suffix_pattern = "")
}
\arguments{
\item{input_file_dir}{directory to scan for bin count files}

\item{suffix_pattern}{suffix to select files from input directory}
}
\value{
data frame with filenames, cell names and file basenames.
}
\description{
Collects all bin count files from given directory
}
