% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varbinGC.R
\name{varbinGC}
\alias{varbinGC}
\title{Compute GC content.}
\usage{
varbinGC(chromFa_dir, data_dir, Nk = "5k")
}
\arguments{
\item{chromFa_dir}{File directory that contains the reference genome.}

\item{data_dir}{File directory. The directory that contains the files of mappable regions, bin boundaries etc.
The GC content file (varbin.gc.content.Nk.txt) will also be output to this directory.}

\item{Nk}{Default value: 5k. It denotes the number of bins. 5000 = 5k, 10000 = 10k, ...}
}
\description{
Given the bin boundaries, compute the GC content in each bin.
}
