% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_clones.R
\name{find_clones}
\alias{find_clones}
\title{Identify clones in hierarchical tree.}
\usage{
find_clones(hc, fdrthresh = -2, sharemin = 0.85, nshare = 3, bymax = T,
  climbfromsize = 2, climbtoshare = 3)
}
\arguments{
\item{hc}{The hclust objects with new items added generated by \code{hclust_tree}.}

\item{fdrthresh}{FDR criterion for clone nodes. Default: -2.}

\item{sharemin}{A feature is considered shared if present in share_min fraction of leaves in a node.Default: 0.90.}

\item{nshare}{Minimal number of shared features in a clone node. Default: 3.}

\item{bymax}{Logical. If TRUE (Default), use maximal of mean FDR for the node to find clones.}

\item{climbfromsize}{An integer.}

\item{climbtoshare}{An integer.}
}
\value{
A hclust object.
}
\description{
Based on hierarchical clustering, identify the hard/soft clones.
}
