% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{sgains_pipeline}
\alias{sgains_pipeline}
\title{Integration with 'sGAINS' tool.}
\usage{
sgains_pipeline(scgv_dir, case_name, varbin_dir, varbin_suffix,
  bins_boundaries_filename, cytoband, badbins = NULL, nsim = 150,
  sharemin = 0.85)
}
\arguments{
\item{scgv_dir}{directory where the results of the analysis should be stored}

\item{case_name}{name of the case to be used for storing results of the analysis}

\item{varbin_dir}{directory where output of 'varbin' step of sGAINS(the binning scheme) is located}

\item{varbin_suffix}{common suffix for files produced by 'varbin' step of sGAINS}

\item{cytoband}{file name for a cytoband coordinate table for the version of the genome being used}

\item{badbins}{a file name for a table of bad bins (bins with outlying read counts) for the specified binning scheme}

\item{nsim}{number of simulations to run for calculating simulated FDR distribution}

\item{sharemin}{a feature is considered 'widely shared' by leaves of a tree node if present in sharemin fraction 
of leaves}

\item{bins_boundaries}{file name for binning scheme to use in the analysis}
}
\description{
This function is called by sGAINS tools to perform the final step in preparation
of results: phylogenetic analysis of single-cell genomes represented by their copy-number profiles
}
